refactor: For better code efficiency and use of less memory in database requests

+++ src/controller/projectController.ts
@@ -4,7 +4,7 @@ import { PrismaClient } from "@prisma/client";
 const prisma = new PrismaClient();
 
 export const createProject = async (req: Request, res: Response) => {
-  const {title, content, tools, linkDeploy, linkRepository, categoryId} = req.body;
+  const {title, content, tools, linkDeploy, linkRepository} = req.body;
   const imageUrl = req.file?.location;
   const userId = req.user?.id;
 
@@ -12,140 +12,157 @@ export const createProject = async (req: Request, res: Response) => {
     res.status(401).json({error: 'Usuário não autenticado.'})
     return;
   }
-  
+
+  const categoryIdInt = parseInt(req.body.categoryId, 10);
+  if(isNaN(categoryIdInt)){
+    res.status(400).json({error: 'Id de categoria inválido'});
+  }
+
+  const projectData = {
+    title: title,
+    content: content,
+    tools: tools,
+    linkDeploy: linkDeploy,
+    linkRepository: linkRepository,
+    categoryId: categoryIdInt,
+    imageUrl: imageUrl,
+    userId: userId
+  }
+
   try{
+    const project = await prisma.projects.create({
+      data: projectData,
+    });
 
-    const categoryIdInt = parseInt(req.body.categoryId, 10);
+    res.status(200).json({message: 'Projeto cadastrado com sucesso!', project});
+  }catch(error: any){
+    console.log('Error ao cadastrar o projeto', error);
 
-
-    const category = await prisma.categories.findUnique({
-      where: {
-        id: categoryIdInt,
-      }
-    })
-
-    if(!category) {
-      res.status(400).json({error: "Categoria não encontrada."});
+    if(error.code === 'P2003') { // Erro do prisma para violação de chave estrangeria
+      res.status(400).json({error: "Categoria não encontrada."})
       return;
     }
 
-    const project = await prisma.projects.create({
-      data: {
-        title: title,
-        content: content,
-        tools: tools,
-        linkDeploy: linkDeploy,
-        linkRepository: linkRepository,
-        categoryId: category.id,
-        imageUrl: imageUrl,
-        userId: userId
-      }
-    });
-
-
-    res.status(200).json({message: 'Projeto cadastrado com sucesso!', project});
-  }catch(error){
-    console.log('Error ao cadastrar o projeto', error);
     res.status(500).json({error: 'Problemas com o servidor. Tente novamente mais tarde!'});
   }
 }
 
 export const getProjects = async (req: Request, res: Response) => {  
-  const {search = '', take = 10, skip = 0} = req.query;
-  
+  const search = String(req.body.search || '').trim();
+  const take = Math.min(Number(req.body.take) || 12, 100);
+  const skip = Math.max(Number(req.body.skip) || 0, 0);
+
   try{
-    const projects = await prisma.projects.findMany({
-      where: {
-        title: {
-          contains: String(search),
-          mode: 'insensitive'
+    //Executar contagem e busca em uma única transação
+    const [projects, totalProjects] = await prisma.$transaction([
+      prisma.projects.findMany({
+        where: {
+          title: {
+            contains: search,
+            mode: 'insensitive'
+          }
         },
-      },
-      take: Number(take),
-      skip: Number(skip)
-    })
+        take,
+        skip
+      }),
+      prisma.projects.count({
+        where: {
+          title: {
+            contains: search,
+            mode: 'insensitive'
+          }
+        }
+      })   
+    ]);
 
-    const totalProjects = await prisma.projects.count({
-      where: {
-        title: {
-          contains: String(search),
-          mode: 'insensitive'
-        },
-      },
-    })
-
-    if(projects.length === 0){
-      res.status(204).json({message: 'Não existem projetos criados.'});
-      return;
-    }
-
-    res.status(200).json({projects: projects, total: totalProjects});
-  }catch(error){ 
-    console.error('Error ao buscar todos os projetos: ' + error);
+    res.status(200).json({projects, totalProjects});
+  }catch(error: any){ 
+    console.error('Error ao buscar todos os projetos: ' + error.message);
     res.status(500).json({error: 'Erro ao buscar projetos. Tente novamente mais tarde!'})
   }
 }
 
 export const getOneProject = async (req: Request, res: Response) => {
-  const { id } = req.params;
+  const projectId = Number(req.params.id);
+
+  if(isNaN(projectId) || projectId <= 0) {
+    res.status(400).json({error: 'Id do projeto inválido.'});
+    return;
+  }
 
   try{
     const project = await prisma.projects.findUnique({
       where: {
-        id: parseInt(id)
-      }
+        id: projectId,
+      },
     });
 
-    if(!project){
-      res.status(404).json({message: 'Projeto não encontrado!'});
+
+    if(!project) {
+      res.status(404).json({error: 'Projeto não encontrado.'});
       return;
     }
 
-    
     res.status(200).json({project});
-  }catch(error){
-    console.error('Error ao buscar o projeto: ' + error);
+  }catch(error: any){
+    console.error('Error ao buscar o projeto: ' + error.message);
     res.status(500).json({error: 'Erro ao buscar o projeto. Tente novamente mais tarde!'})
   }
 }
 
 export const getUserProject = async (req: Request, res: Response) => {
   const { username } = req.query;
+
+  if(!username || typeof username !== 'string') {
+    res.status(400).json({error: 'Username inválido ou não fornecido'});
+    return;
+  }
   
   try{
     const user = await prisma.users.findUnique({
-      where: {
-        username: String(username)
-      }
-    });
+      where: {username: String(username)},
+      select: {id: true} // Selecionando apenas o ID   
+    })
 
+    if(!user) {
+      res.status(400).json({error: 'Usuário não encontrado'});
+      return;
+    }
     
     const projects = await prisma.projects.findMany({
       where: {
-        userId: user?.id
-      }
+        userId: user.id,
+      },
+      select: {id: true, title: true, tools: true}, 
     })
 
-    if(!projects) {
-      res.status(404).json({message: 'Projetos não encontrados!'})
+
+    if(projects.length === 0) {
+      res.status(404).json({message: 'Projetos não encontrados.'})
+      return;
     }
 
     res.status(200).json({projects})
-  }catch(error){
-    console.error('Error ao buscar o projeto: ' + error);
-    res.status(500).json({error: 'Erro ao buscar o projeto. Tente novamente mais tarde!'})
+  }catch(error: any){
+    console.error('Error ao buscar o projeto: ' + error.message);
+    res.status(500).json({error: 'Erro ao buscar o projeto. Tente novamente mais tarde!'});
   }
 }
 
 export const updateProject = async (req: Request, res: Response) => {
-  const {id} = req.params;
+  const id = Number(req.params.id);
   const {title, content, tools, linkDeploy, linkRepository} = req.body;
   const userId = req.user?.id;
 
+  if(!userId) {
+    res.status(401).json({message: 'Usuário não autenticado'});
+    return;
+  }
+ 
   try{
     const existProject = await prisma.projects.findUnique({
       where:{
-        id: parseInt(id)
+        id: id
       }
     });
 
@@ -161,7 +178,7 @@ export const updateProject = async (req: Request, res: Response) => {
 
     const updateProject = await prisma.projects.update({
       where:{
-        id: parseInt(id)
+        id: id
       },
       data: {
         title: title,
@@ -173,25 +190,30 @@ export const updateProject = async (req: Request, res: Response) => {
       }});
 
     res.status(200).json({message: 'Projeto atualizado com sucesso.', project: updateProject})
-  }catch(error){
-    console.error('Error ao atualizar projeto: ' + error);
-    res.status(500).json({error: 'Erro ao atualizar projeto. Tente novamente mais tarde!'})
+  }catch(error: any){
+    console.error('Error ao atualizar projeto: ' + error.message);
+    res.status(500).json({error: 'Erro ao atualizar projeto. Tente novamente mais tarde!'});
   }
 }
 
 export const deleteProject = async (req: Request, res: Response) => {
-  const { id } = req.params;
+  const id = Number(req.params.id);
   
+  if(isNaN(id)) {
+    res.status(400).json({error: 'Id inválido. Deve ser um número.'});
+    return;
+  }
+
   try{
     await prisma.projects.delete({
       where: {
-        id: parseInt(id)
+        id: id
       }
     })
     
     res.status(200).json({message: 'Projeto removido com sucesso!'});
-  }catch(error){
-    console.error('Error ao excluir projeto: ' + error);
+  }catch(error: any){
+    console.error('Error ao excluir projeto: ' + error.message);
     res.status(500).json({error: 'Erro ao excluir projeto. Tente novamente mais tarde!'})
   }
 }
